🔧 循環依存を解決してVitest実行可能に修正
- GameState enumを独立したtypes/GameState.tsに分離
- 更新履歴定数をconstants/ChangelogConstants.tsに分離
- BaseOutGameScene用のIGameContextインターフェースを作成
- 循環依存チェーンを断ち切り、npm run vitestが正常動作
