+++ src/game/constants/ChangelogConstants.ts
@@ -1,0 +1,10 @@
+/**
+ * 更新履歴関連の定数
+ */
+export class ChangelogConstants {
+    /** 更新履歴が未表示であることを示す特別な値 */
+    public static readonly CHANGELOG_INDEX_NONE = -1;
+    
+    /** 初期値（更新履歴モーダルが表示されない） */
+    public static readonly CHANGELOG_INDEX_INITIAL = -2;
+}
